microservices/validator-service/requirements.txt
# 验证服务依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# 数据库和缓存
redis==5.0.1

# 配置文件处理
pyyaml==6.0.1

# 监控
prometheus-client==0.19.0

# 日志
structlog==23.2.0